Add argument checking to first pass
This involves checking for valid numeric constants in argument positions
as well as that all attempted symbol usage is resolvable.

This also exposed a bug (since fixed) in processing the LIT directive
which caused literals used before their definitions to show up twice in
the symbols table, and ultimately get reported as being undefined.