Add `RPT` and `PAD` processor directives
RPT <N> causes the assembler to behave as if it saw the remainder of the line
N times at that location.

PAD <M> adds whichever command is given by the name #define-d to
`CORE_ERR_STR` to the assembled output enough times so that the command
following it is aligned to an M-byte boundary.

The also caused the addition of grabCmd() to handle loading cmdStr, since
RPT needs to get the real command string in the course setting up the
repetition.